🖌️UI : 디자인 수정, Quest 진행률, Boss HP UI 추가

+++ Netherveil/Source/Netherveil/Public/Enemy/EnemyFSM.h
@@ -70,7 +70,7 @@ public:
 	virtual void OnDamageProcess();
 
 	UPROPERTY(EditDefaultsOnly, BlueprintReadWrite, Category = FSM)
-	int32 hp = 3;
+	float hp = 3.0;
 
 	UPROPERTY(EditAnywhere, Category = FSM)
 	float damageDelayTime = 2.0f;
